LPS-196086 refactor IndividualSegmentsChecker.java
moving class to CheckIndividualSegmentsSchedulerJobConfiguration.java and deleting the other one
